+++ src/main/java/evg/testt/model/AuthorshipPoints.java
@@ -3,26 +3,14 @@ package evg.testt.model;
 import javax.persistence.*;
 
 
-@Entity(name = "UserPoints")
-public class UserAuthorshipPoints extends BaseModel {
+@Entity(name = "authorshipPoints")
+public class AuthorshipPoints extends BaseModel {
 
     @OneToOne(cascade = CascadeType.MERGE )
     @JoinColumn(name = "user_id")
     private User user;
 
-
-    private Integer knowledgeLevelPoints;
-
     private Integer authorshipPoints;
-
-
-    public Integer getKnowledgeLevelPoints() {
-        return knowledgeLevelPoints;
-    }
-
-    public void setKnowledgeLevelPoints(Integer knowledgeLevelPoints) {
-        this.knowledgeLevelPoints = knowledgeLevelPoints;
-    }
 
     public Integer getAuthorshipPoints() {
         return authorshipPoints;
@@ -40,12 +28,6 @@ public class UserAuthorshipPoints extends BaseModel {
         this.user = user;
     }
 
-    public void increaseKnowledgeLevel(Integer increment) {
-        knowledgeLevelPoints += increment;
-    }
-    public void decreaseKnowledgeLevel(Integer decrement) {
-        knowledgeLevelPoints -= decrement;
-    }
 
     public void increaseAuthorshipPoints(Integer increment) {
         authorshipPoints += increment;
