+++ src/main/java/evg/testt/service/AuthorshipPointsService.java
@@ -1,0 +1,6 @@
+package evg.testt.service;
+
+import evg.testt.model.AuthorshipPoints;
+
+public interface AuthorshipPointsService extends Service<AuthorshipPoints>{
+}
