+++ src/main/java/evg/testt/service/KnowledgeLevelPointsService.java
@@ -1,0 +1,6 @@
+package evg.testt.service;
+
+import evg.testt.model.KnowledgeLevelPoints;
+
+public interface KnowledgeLevelPointsService extends Service<KnowledgeLevelPoints> {
+}
