--- src/main/java/evg/testt/service/UserPointsService.java
@@ -1,6 +1,0 @@
-package evg.testt.service;
-
-import evg.testt.model.UserPoints;
-
-public interface UserPointsService extends Service<UserPoints>{
-}
