+++ src/main/java/evg/testt/service/impl/AuthorshipPointsServiceImpl.java
@@ -1,0 +1,10 @@
+package evg.testt.service.impl;
+
+import evg.testt.dao.AuthorshipPointsDao;
+import evg.testt.model.AuthorshipPoints;
+import evg.testt.service.AuthorshipPointsService;
+import org.springframework.stereotype.Service;
+
+@Service
+public class AuthorshipPointsServiceImpl extends BaseService<AuthorshipPoints, AuthorshipPointsDao> implements AuthorshipPointsService {
+}
