+++ src/main/java/evg/testt/service/impl/KnowledgeLevelPointsServiceImpl.java
@@ -1,0 +1,11 @@
+package evg.testt.service.impl;
+
+
+import evg.testt.dao.KnowledgeLevelPointsDao;
+import evg.testt.model.KnowledgeLevelPoints;
+import evg.testt.service.KnowledgeLevelPointsService;
+import org.springframework.stereotype.Service;
+
+@Service
+public class KnowledgeLevelPointsServiceImpl extends BaseService<KnowledgeLevelPoints, KnowledgeLevelPointsDao> implements KnowledgeLevelPointsService {
+}
